Minor changes to home page and attached Things to Do list
Made some minor adjustments to the home page in the CSS and HTML files.
Modified the Name placeholder text from "First Name" to "Name".
Added a "Like us on Facebook" picture to be used as a link to the BRO
facebook page.
Added a progress update file of the Things to Do and feedback from
usability testers.