Modified home page, events page, announcements page and people page.
Made changes in accordance with Progress Update file (Things to Do word
document).
Changed the CSS for the events, announcements and people page to have
the same layout as home page.
Increased the orchestra photo and  next concert photo on the home page.
Added the 2017 concerts program and Queens Gardens concert photos to the
events page.
Set up the layout for the musicians on display in the people.html file.